Created Entites, Cells, and Movement
Created an entity handler and a little test food class that really does nothing but draw a square. Then I made a cell class that allows basic tweening movement to a specified target point. I dont know if I like the way tweening looks currently and considering that the distance between the current position and the target point could vary, so does the speed. I may ultimately end up changing it to a physics based movement style with velocities and friction. We'll see when movement becomes more necessary.